controller 추가

+++ spingTest/src/main/java/org/zerock/service/BoardService.java
@@ -7,11 +7,11 @@ import org.zerock.domain.BoardVO;
 public interface BoardService {
 	public void register(BoardVO board);
 	
-	public BoardVO get(Long bno);
+	public BoardVO get(int bno);
 	
 	public boolean modify(BoardVO board);
 	
-	public boolean remove(Long bno);
+	public boolean remove(int bno);
 	
 	public List<BoardVO> getList();
 }
